Add files via upload
Added a background for the game. I also added more text to the game to guide the user. I am trying to make the actual battle part of the game work, still coding the part where each user can press keys on the keyboard to attack the enemy with its respective attack.